Promote was being triggered twice.

+++ scripts/delete-pv.sh
@@ -4,6 +4,7 @@ context=$1
 namespace=$2
 networkSuffix=$3
 network=$4
+deployEnv=$5
 
 echo "delete-pv.sh"
 
@@ -11,6 +12,7 @@ echo "context: ${context}"
 echo "namespace: ${namespace}"
 echo "networkSuffix: ${networkSuffix}"
 echo "network: ${network}"
+echo "deployEnv: ${deployEnv}"
 
 kubeContextArg=""
 if [[ ${context} != "" ]]
@@ -25,9 +27,13 @@ then
 fi
 
 cat ./lightning-kube-pvc.yaml | sed "s/\X_NETWORK_SUFFIX_X/${networkSuffix}/" | sed "s/\X_NETWORK_X/${network}/" | kubectl ${kubeContextArg} ${namespaceArg} delete -f -
-cat ./lightning-kube-pv.yaml | sed "s/\X_NETWORK_SUFFIX_X/${networkSuffix}/" | kubectl ${kubeContextArg} ${namespaceArg} delete -f -
+
+if [[ ${deployEnv} != "gke" ]]
+then
+    cat ./lightning-kube-pv.yaml | sed "s/\X_NETWORK_SUFFIX_X/${networkSuffix}/" | kubectl ${kubeContextArg} ${namespaceArg} delete -f -
+fi
 
 #cat ./bitcoind-kube-pvc.yaml | sed "s/\X_NETWORK_SUFFIX_X/${networkSuffix}/" | kubectl ${kubeContextArg} ${namespaceArg} delete -f -
 
 
-#./delete-pv.sh "" lightning-kube-simnet -simnet simnet
+#./delete-pv.sh "" lightning-kube-simnet -simnet simnet gke
